Rename HtmlRootComponent -> HtmlComponent
