+++ src/Components/Web/src/HtmlRendering/HtmlRenderer.cs
@@ -43,7 +43,7 @@ public sealed class HtmlRenderer : IDisposable, IAsyncDisposable
     /// <summary>
     /// Adds an instance of the specified component and instructs it to render. The resulting content represents the
     /// initial synchronous rendering output, which may later change. To wait for the component hierarchy to complete
-    /// any asynchronous operations such as loading, await <see cref="HtmlRootComponent.QuiescenceTask"/> before
+    /// any asynchronous operations such as loading, await <see cref="HtmlComponent.QuiescenceTask"/> before
     /// reading content from the <see cref="HtmlComponent"/>.
     /// </summary>
     /// <typeparam name="TComponent">The component type.</typeparam>
@@ -54,7 +54,7 @@ public sealed class HtmlRenderer : IDisposable, IAsyncDisposable
     /// <summary>
     /// Adds an instance of the specified component and instructs it to render. The resulting content represents the
     /// initial synchronous rendering output, which may later change. To wait for the component hierarchy to complete
-    /// any asynchronous operations such as loading, await <see cref="HtmlRootComponent.QuiescenceTask"/> before
+    /// any asynchronous operations such as loading, await <see cref="HtmlComponent.QuiescenceTask"/> before
     /// reading content from the <see cref="HtmlComponent"/>.
     /// </summary>
     /// <typeparam name="TComponent">The component type.</typeparam>
@@ -67,7 +67,7 @@ public sealed class HtmlRenderer : IDisposable, IAsyncDisposable
     /// <summary>
     /// Adds an instance of the specified component and instructs it to render. The resulting content represents the
     /// initial synchronous rendering output, which may later change. To wait for the component hierarchy to complete
-    /// any asynchronous operations such as loading, await <see cref="HtmlRootComponent.QuiescenceTask"/> before
+    /// any asynchronous operations such as loading, await <see cref="HtmlComponent.QuiescenceTask"/> before
     /// reading content from the <see cref="HtmlComponent"/>.
     /// </summary>
     /// <param name="componentType">The component type. This must implement <see cref="IComponent"/>.</param>
@@ -79,13 +79,13 @@ public sealed class HtmlRenderer : IDisposable, IAsyncDisposable
     /// <summary>
     /// Adds an instance of the specified component and instructs it to render. The resulting content represents the
     /// initial synchronous rendering output, which may later change. To wait for the component hierarchy to complete
-    /// any asynchronous operations such as loading, await <see cref="HtmlRootComponent.QuiescenceTask"/> before
-    /// reading content from the <see cref="HtmlRootComponent"/>.
+    /// any asynchronous operations such as loading, await <see cref="HtmlComponent.QuiescenceTask"/> before
+    /// reading content from the <see cref="HtmlComponent"/>.
     /// </summary>
     /// <param name="componentType">The component type. This must implement <see cref="IComponent"/>.</param>
     /// <param name="parameters">Parameters for the component.</param>
     /// <returns>An <see cref="HtmlComponent"/> instance representing the render output.</returns>
-    public HtmlRootComponent BeginRenderingComponent(
+    public HtmlComponent BeginRenderingComponent(
         [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)] Type componentType,
         ParameterView parameters)
         => _passiveHtmlRenderer.BeginRenderingComponent(componentType, parameters);
